fix: 干净克隆全量重建暴露的 26 处告警清零（含 3 处真缺陷）
上一轮「0 告警」是假象：本地 build 是增量的，已编译过的对象不再重报告警。
本轮从 GitHub 全新克隆 + touch 全部自有源码强制全量重建，重新暴露 26 处
（克隆 22 处 + 仅 client=ON 才编译的 4 处），全部修完。

真缺陷（不只是告警）：
- engine/net/src/tcp_transport.cpp、client/core/src/net_link.cpp：
  FIONBIO 是 unsigned long 常量 0x8004667E，ioctlsocket 形参却是 long，
  隐式转换把 2147772030 变成 -2147195266。显式 static_cast 表明按位传值是有意的。
- server/gamenode/combat/src/buff/buff_system.cpp：
  BuffRemoved.stacks 是 uint8_t、BuffInstance.stacks 是 uint16_t，
  原赋值在层数 >255 时会静默回绕。改为饱和截断到 255。

其余为类型/未使用类告警：
- buff_system.cpp：exist_idx 由 int(-1 哨兵) 改 size_t 哨兵，消除 6 处
  int→size_type 符号转换；vec.erase 下标显式转 ptrdiff_t；显式丢弃 Apply 的
  未用参数 trace（BuffApplied 事件体受 static_assert(sizeof<=32) 限制无 trace 字段）。
- engine/rpc/src/grpc_channel_pool.cpp：round-robin 取模结果 size_t→uint32_t 显式转换。
- client/core/src/net_link.cpp：mode 变量遮蔽（-Wshadow），用块作用域隔开。
- client/renderer/src/renderer.cpp：volatile sink（防循环被优化掉）显式 (void) 消费。
- engine/net/tests/net_test.cpp：12 处忽略 nodiscard 的 Poll 返回值，补 (void)。

验证：
- 本地（MMORPG_BUILD_CLIENT=ON）全量重建 0 error / 0 warning；ctest 49/49 通过。
- 干净克隆（client 默认 OFF）全量构建 0 error；ctest 46/46 通过。
  （少 3 条是 Client/Renderer/Resource 三个 Godot 客户端测试，依赖未入 git 的
  godot-cpp 嵌套仓库，克隆环境默认关闭，非回归。）
- 运行期：5 个 daemon --run-for 全部 exit 0 且 stderr 为空；
  gateway e2e SMOKE_OK（auth_ok=1、frames_in=4、断开后 Suspended）；
  部署包解压全新目录 smoke 全绿 + start/stop 优雅停止往返通过。

+++ engine/rpc/src/grpc_channel_pool.cpp
@@ -47,7 +47,8 @@ Result<std::shared_ptr<grpc::Channel>> GrpcChannelPool::Get(std::string_view tar
         (void)ch->WaitForConnected(std::chrono::system_clock::now() + std::chrono::milliseconds{100});
     }
     auto out = ch;
-    e.next = (e.next + 1) % per_target_;
+    // % 的操作数是 size_t，结果回落给 uint32_t 必须显式转换（-Wconversion）
+    e.next = static_cast<uint32_t>((e.next + 1) % per_target_);
     return Result<std::shared_ptr<grpc::Channel>>::Ok(std::move(out));
 }
